Revert "Merge commit '3fff4c532c363ac933c9'"
This reverts commit fb920a3d44870a4b3d44c41757838b6bdb0d8ba0, reversing
changes made to 5f711c88536b7e5304e8faa8bd496541c5feb3b6.

+++ dialogs/DebugWorldInputDlg.cpp
@@ -208,7 +208,7 @@ CInsertUnicodeDlg dlg;
 
 unsigned char utf8 [10];
 
-int iLen = _pcre_ord2utf8 ((int) dlg.m_iCode, utf8);
+int iLen = _pcre_ord2utf8 (dlg.m_iCode, utf8);
 
 
 CString strText;
